requirements.txt
Pillow==9.5.0
layoutparser
torchvision
git+https://github.com/facebookresearch/detectron2.git@v0.5#egg=detectron2
pytesseract
pymupdf
pymupdf-fonts
pydantic
pydantic-settings
nougat
transformers
scikit-learn